requirements.txt
feedparser>=6.0
httpx>=0.27
pyyaml>=6.0
rich>=13.0
flask>=3.0

